# packages/generated/api-client-csharp/api/openapi.yaml
openapi: 3.1.0
info:
  title: Plerion
  version: 0.1.0
servers:
- url: /
security:
- oauth2:
  - openid
- bearerAuth: []
paths:
...
  /groups:
    delete:
      operationId: delete_groups
      parameters:
      - description: List of Ids to delete
        explode: true
        in: query
        name: ids
        required: true
        schema:
          description: List of Ids to delete
          items:
            format: uuid
            type: string
            x-schema-name: delete_groups_param_ids_item
            nullable: false
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Delete Groups
      tags:
      - Default
    get:
      operationId: get_groups
      parameters:
      - description: Optional list of Ids to filter by
        explode: true
        in: query
        name: ids
        required: false
        schema:
          items:
            format: uuid
            type: string
            x-schema-name: get_groups_param_ids_item
            nullable: false
          type: array
          nullable: true
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/GroupRead"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Groups
      tags:
      - Default
    patch:
      operationId: update_groups
      parameters:
      - explode: true
        in: query
        name: allow_missing
        required: false
        schema:
          default: false
          title: Allow Missing
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: "#/components/schemas/GroupBatchUpdate"
              type: array
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/GroupRead"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Update Groups
      tags:
      - Default
    post:
      operationId: create_group
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupCreate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupRead"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Create Group
      tags:
      - Default
  /layers:
    delete:
      operationId: delete_layers
      parameters:
      - description: List of Ids to delete
        explode: true
        in: query
        name: ids
        required: true
        schema:
          description: List of Ids to delete
          items:
            format: uuid
            type: string
            x-schema-name: delete_layers_param_ids_item
            nullable: false
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Delete Layers
      tags:
      - Default
    get:
      operationId: get_layers
      parameters:
      - description: Optional list of Ids to filter by
        explode: true
        in: query
        name: ids
        required: false
        schema:
          items:
            format: uuid
            type: string
            x-schema-name: get_layers_param_ids_item
            nullable: false
          type: array
          nullable: true
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/LayerRead"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Layers
      tags:
      - Default
    patch:
      operationId: update_layers
      parameters:
      - explode: true
        in: query
        name: allow_missing
        required: false
        schema:
          default: false
          title: Allow Missing
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: "#/components/schemas/LayerBatchUpdate"
              type: array
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/LayerRead"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Update Layers
      tags:
      - Default
    post:
      operationId: create_layer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LayerCreate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LayerRead"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Create Layer
      tags:
      - Default
...
components:
  schemas:
...
    GroupBatchUpdate:
      example:
        parent_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          title: Id
          type: string
        name:
          nullable: true
          type: string
        parent_id:
          format: uuid
          nullable: true
          type: string
      required:
      - id
      title: GroupBatchUpdate
    GroupCreate:
      example:
        parent_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          nullable: true
          type: string
        name:
          title: Name
          type: string
        parent_id:
          format: uuid
          nullable: true
          type: string
      required:
      - name
      title: GroupCreate
    GroupRead:
      example:
        updated_at: 2000-01-23T04:56:07.000+00:00
        parent_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          title: Id
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
        name:
          title: Name
          type: string
        parent_id:
          format: uuid
          nullable: true
          type: string
      required:
      - created_at
      - id
      - name
      - updated_at
      title: GroupRead
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
      title: HTTPValidationError
    LayerBatchUpdate:
      example:
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          title: Id
          type: string
        name:
          nullable: true
          type: string
      required:
      - id
      title: LayerBatchUpdate
    LayerCreate:
      example:
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          nullable: true
          type: string
        name:
          title: Name
          type: string
      required:
      - name
      title: LayerCreate
    LayerRead:
      example:
        updated_at: 2000-01-23T04:56:07.000+00:00
        name: name
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          title: Id
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
        name:
          title: Name
          type: string
      required:
      - created_at
      - id
      - name
      - updated_at
      title: LayerRead
...
    ValidationError:
      example:
        msg: msg
        loc:
        - ValidationError_loc_inner
        - ValidationError_loc_inner
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/ValidationError_loc_inner"
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ValidationError_loc_inner:
      anyOf:
      - type: string
      - type: integer
  securitySchemes: {}